+++ app/models/charms/essence_charm.rb
@@ -6,5 +6,16 @@ module Charms
     def charm_type
       'Essence'
     end
+
+    def self.from_charm!(charm)
+      return charm if charm.type == 'Charms::EssenceCharm'
+
+      new_charm = charm.becomes(Charms::EssenceCharm)
+      new_charm.type = 'Charms::EssenceCharm'
+      new_charm.ability = ''
+      new_charm.min_ability = nil
+      new_charm.save!
+      new_charm
+    end
   end
 end