+++ app/models/custom_essence_character.rb
@@ -8,4 +8,18 @@ class CustomEssenceCharacter < Character
   attribute :exalt_type, :string, default: 'Essence Exalt'
 
   alias_attribute :charms, :essence_charms
+
+  def self.from_character!(character)
+    new_cha = character.becomes(CustomEssenceCharacter)
+    new_cha.type = 'CustomEssenceCharacter'
+    new_cha.caste_attributes = []
+    new_cha.favored_attributes = []
+    new_cha.favored_abilities = []
+    new_cha.supernal_ability = nil
+    new_cha.save!
+    (new_cha.attribute_charms + new_cha.ability_charms).each do |charm|
+      EssenceCharm.from_charm!(charm)
+    end
+    new_cha
+  end
 end