+++ app/serializers/dragonblood_character_serializer.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-# app/serializers/dragonblood_character_serializer.rb
-class DragonbloodCharacterSerializer < ExaltSerializer
-  attributes :caste_abilities, :favored_abilities, :aura
-end